+++ libs/graphs/include/mrpt/graphs/dijkstra.h
@@ -465,6 +465,14 @@ class CDijkstra
 
 	}  // end getTreeGraph
 
+	/// \overload \note (New in MRPT 2.2.1)
+	tree_graph_t getTreeGraph() const
+	{
+		tree_graph_t tg;
+		getTreeGraph(tg);
+		return tg;
+	}
+
 	/** @} */
 
 };	// end class